requirements.txt
dm-haiku
optax
jax